move 'handlers' into their own module
